refactor: establish config/ and launcher/ modules with initial file moves
