+++ panel-plugin/config/usage-stats.h
