+++ panel-plugin/launcher/category-button.cpp
