+++ panel-plugin/launcher/category-button.h
