+++ panel-plugin/launcher/category.cpp
