+++ panel-plugin/launcher/category.h
